+++ gcc/testsuite/g++.target/riscv/tt/sfpi/lut-select-refuse-arity-bh.C
@@ -5,9 +5,9 @@
 // { dg-final { scan-tree-dump-not "formed " "rvtt_lut_select" } }
 // { dg-final { scan-assembler-not "SFPLUTFP32" } }
 
-extern volatile unsigned __instrn_buffer[];
+extern volatile unsigned long __instrn_buffer[];
 namespace ckernel {
-constexpr inline volatile unsigned (&instrn_buffer)[] = ::__instrn_buffer;
+constexpr inline volatile unsigned long (&instrn_buffer)[] = ::__instrn_buffer;
 }
 #include <sfpi.h>
